Move omission struct to structs folder
