+++ src/csb/examination/pages/candidate.rs
@@ -4,16 +4,14 @@ use axum::response::{IntoResponse, Response};
 use crate::{
     AppError, Context, CsbContext, CsbStore, ElectoralDistrict, HtmlTemplate,
     candidate_lists::CandidateListId,
-    csb::{
-        Omission,
-        examination::{
-            extractors::CsbPoliticalGroup,
-            pages::CsbCandidatePath,
-            structs::{PaperCorrected, PaperCorrectedPersonDetails},
-        },
+    csb::examination::{
+        extractors::CsbPoliticalGroup,
+        pages::CsbCandidatePath,
+        structs::{PaperCorrected, PaperCorrectedPersonDetails},
     },
     filters,
     persons::Person,
+    structs::csb::Omission,
 };
 
 #[derive(Template)]
@@ -93,6 +91,7 @@ mod tests {
 
     use crate::{
         persons::PersonId,
+        structs::csb::OmissionCategory,
         test_utils::{response_body_string, sample_candidate_list, sample_person},
     };
 
@@ -287,8 +286,6 @@ mod tests {
 
     #[tokio::test]
     async fn renders_added_candidate_omissions_as_badges() {
-        use crate::csb::OmissionCategory;
-
         let store = CsbStore::new_for_test();
         let stream_id = store.stream_id;
