+++ src/csb/examination/pages/omission/views.rs
@@ -4,13 +4,11 @@ use axum_extra::routing::TypedPath;
 use crate::{
     AppError, Context, CsbStore, ElectoralDistrict, Locale, Overlay, QueryParamState,
     candidate_lists::CandidateListId,
-    csb::{
-        Omission, OmissionPlaceholders, OmissionType,
-        examination::{OmissionForm, pages::CsbDeleteOmissionPath},
-    },
+    csb::examination::{OmissionForm, pages::CsbDeleteOmissionPath},
     filters,
     form::FormData,
     persons::PersonId,
+    structs::csb::{Omission, OmissionPlaceholders, OmissionType},
 };
 
 use super::OmissionTarget;
