+++ src/csb/mod.rs
@@ -11,12 +11,10 @@ pub mod index;
 pub mod monitoring;
 
 mod context;
-mod omission;
 mod store_csb;
 mod store_main;
 
 pub use context::CsbContext;
-pub use omission::{Omission, OmissionCategory, OmissionId, OmissionPlaceholders, OmissionType};
 pub use store_csb::{CsbEvent, CsbStoreData};
 #[cfg(any(test, feature = "dev-features"))]
 pub use store_main::CsbMainEvent;
