+++ src/structs/csb/mod.rs
@@ -1,0 +1,6 @@
+mod omission;
+
+pub use omission::{Omission, OmissionCategory, OmissionId, OmissionPlaceholders, OmissionType};
+
+#[cfg(test)]
+pub use omission::tests::sample_omission;
