+++ src/structs/mod.rs
@@ -1,1 +1,2 @@
+pub mod csb;
 pub mod typst;
